LPS-79881 The uuid for these classes and the verification were added between 6.1 and prior to 6.2 so we need to execute the verification only in the upgrade process to 6.2
