+++ modules/apps/collaboration/message-boards/message-boards-service/src/main/java/com/liferay/message/boards/internal/verify/model/MBDiscussionVerifiableModel.java
@@ -22,8 +22,7 @@ import com.liferay.portal.kernel.verify.model.VerifiableUUIDModel;
  * @author Miguel Pastor
  */
 public class MBDiscussionVerifiableModel
-	implements VerifiableAuditedModel, VerifiableGroupedModel,
-			   VerifiableUUIDModel {
+	implements VerifiableAuditedModel, VerifiableGroupedModel {
 
 	@Override
 	public String getJoinByTableName() {
