+++ portal-impl/src/com/liferay/portal/verify/model/MBThreadVerifiableModel.java
@@ -12,24 +12,24 @@
  * details.
  */
 
-package com.liferay.journal.internal.verify.model;
+package com.liferay.portal.verify.model;
 
+import com.liferay.portal.kernel.verify.model.VerifiableAuditedModel;
 import com.liferay.portal.kernel.verify.model.VerifiableUUIDModel;
 
 /**
  * @author Miguel Pastor
  */
-public class JournalArticleResourceVerifiableModel
-	implements VerifiableUUIDModel {
+public class MBThreadVerifiableModel implements VerifiableUUIDModel {
 
 	@Override
 	public String getPrimaryKeyColumnName() {
-		return "resourcePrimKey";
+		return "threadId";
 	}
 
 	@Override
 	public String getTableName() {
-		return "JournalArticleResource";
+		return "MBThread";
 	}
 
 }
