+++ portal-impl/src/com/liferay/portal/verify/model/PollsVoteVerifiableModel.java
@@ -12,7 +12,7 @@
  * details.
  */
 
-package com.liferay.polls.internal.verify.model;
+package com.liferay.portal.verify.model;
 
 import com.liferay.portal.kernel.verify.model.VerifiableAuditedModel;
 import com.liferay.portal.kernel.verify.model.VerifiableGroupedModel;
@@ -21,13 +21,7 @@ import com.liferay.portal.kernel.verify.model.VerifiableUUIDModel;
 /**
  * @author Miguel Pastor
  */
-public class PollsVoteVerifiableModel
-	implements VerifiableAuditedModel, VerifiableGroupedModel {
-
-	@Override
-	public String getJoinByTableName() {
-		return "questionId";
-	}
+public class PollsVoteVerifiableModel implements VerifiableUUIDModel {
 
 	@Override
 	public String getPrimaryKeyColumnName() {
@@ -35,38 +29,8 @@ public class PollsVoteVerifiableModel
 	}
 
 	@Override
-	public String getRelatedModelName() {
-		return "PollsQuestion";
-	}
-
-	@Override
-	public String getRelatedPKColumnName() {
-		return "questionId";
-	}
-
-	@Override
-	public String getRelatedPrimaryKeyColumnName() {
-		return "questionId";
-	}
-
-	@Override
-	public String getRelatedTableName() {
-		return "PollsQuestion";
-	}
-
-	@Override
 	public String getTableName() {
 		return "PollsVote";
-	}
-
-	@Override
-	public boolean isAnonymousUserAllowed() {
-		return true;
-	}
-
-	@Override
-	public boolean isUpdateDates() {
-		return false;
 	}
 
 }
